Move parse and parser_for_xml to Feedjira top level module
This substantially cleans up the API for parsing
