+++ lib/feedjira.rb
@@ -43,4 +43,32 @@ module Feedjira
   class NoParserAvailable < StandardError; end
 
   extend Configuration
+
+  # Parse XML with first compatible parser
+  #
+  # @example
+  #   xml = HTTParty.get("http://example.com").body
+  #   Feedjira.parse(xml)
+  def parse(xml, &block)
+    parser = parser_for_xml(xml)
+
+    if parser.nil?
+      raise NoParserAvailable, "No valid parser for XML."
+    end
+
+    parser.parse xml, &block
+  end
+  module_function :parse
+
+  # Find compatible parser for given XML
+  #
+  # @example
+  #   xml = HTTParty.get("http://example.com").body
+  #   parser = Feedjira.parser_for_xml(xml)
+  #   parser.parse(xml)
+  def parser_for_xml(xml)
+    start_of_doc = xml.slice(0, 2000)
+    Feedjira.parsers.detect { |klass| klass.able_to_parse?(start_of_doc) }
+  end
+  module_function :parser_for_xml
 end
